Delay space allocation until after (pre)stage
ENDIT relies on space set aside for (pre)stage, and as such doesn't need
space allocated in the pool until the file is available.

Delaying allocation should enable ENDIT to see the entire request queue
when we're dealing with large recalls, and is needed to cope with future
tape technology where tape sizes are many times larger than the
available fast staging space on a tape pool host.

Credits go to Xavier Mol for a very detailed description on how to
accomplish this.